+++ packages/extension/src/views/Events/EventPreview/InitializedPreview.tsx
@@ -5,7 +5,7 @@ import { Link } from './components'
 export function InitializedPreview() {
   return (
     <>
-      <Text>useDApp was detected on the page and the DevTools extension was initialized.</Text>
+      <Text>useDAppify was detected on the page and the DevTools extension was initialized.</Text>
       <Link block href="https://usedapp.readthedocs.io/en/latest/">
         Read the official documentation
       </Link>