+++ packages/extension/src/views/Events/EventPreview/NetworkDisconnectedPreview.tsx
@@ -2,5 +2,5 @@ import React from 'react'
 import { Text } from '../../shared'
 
 export function NetworkDisconnectedPreview() {
-  return <Text>Network disconnected. useDApp will not make network calls.</Text>
+  return <Text>Network disconnected. useDAppify will not make network calls.</Text>
 }